Allow custom QueueManagers, export StompServer to other modules, other fixes.
Don't throw StompFrames as exceptions, because the handler that forwarded
them all to the client broke when other types of exceptions occurred. Simpler
to use error return values here instead.

Create QueueManagers per connection using a class given as a new argument
to the StompServer and SecureStompServer constructors.

Don't immediately start listening. Export symbols so that the calling module
has a chance to override QueueManager, etc. Give a usage example near the
top of the file.