Fix SSL issues due to small API differences between TLS streams and Net sockets.
Secure and non-secure STOMP connections share the same code. But due to
UTF-8 conversion, the stream API used for secure connections silently
drops the NUL character that serves as a frame terminator while
sending, causing the other side to lump the remainder of the stream
into the body of the first frame.

The parameter to net.createServer's callback gets a 'connect' event,
while tls.createServer's equivalent has already connected, so it
doesn't get any kind of 'connect' or 'secureConnection' event.